Improved bootstrap process
New bootstrap system which is flexible and allows for additional
bootstrap setups, for instance a base setup, ez publish tweaks and
prod/dev configurations.
The new system creates a config and an app instance which initialized
from the configuration files and enviroment variables. The resulting
config and bootstrap process can be cached in `build` for production
modes.