Improve logging by using monolog for the backend
The error handler now no longer hard codes the eZ debug calls. Instead
it now support setting a logger, if a logger is set it will then send
the error report to the logger.
In addition the error handler will now log errors separately from
reporting errors, this means that an error may be logged and then
reported.

The base app now has a system for setting up loggers and log handlers,
each logger defines a channel and the attached handlers. The handlers
define how log records are handled.
The default setup now creates a handler which forwards errors to
eZ debug when the ezp mode is active.