Converted the sentry logging to use the new log handlers instead
Sentry is now just a regular handler which will be enabled when a
DSN value is set.

The setup of the ezp error handler/logger has also changed, now it
always installs an error handler and logger. The exception is if
app.errorMode or `STARTER_ERROR_MODE` is set to null.

Added support for a setup callback for log handlers, if set it will
call this with definition array and the callback must instantiate
the handler for us.
Level and bubble are now also read from the definition and passed to
the handlers.