Improved support for editors in pretty page handler
The editor is now properly setup from editor.name and
editor.editors. If an editor is chosen a callback is used to
proplery map remote paths to local paths, this all depends on
editor.remoteFileSystem and editor.fileMappings.

An editor may also contain an ajax: prefix which means the url
is opened using ajax.