Replaced use of WhoopsRun with Run from Whoops
The WhoopsRun class was a copy of Run with some types removed
for compatibility with PHP 7. However this caused Whoops 1
to fail as the type declarations for setRun required a Run
instance and WhoopsRun did not inherit (and cannot).

As the Whoops 2 code uses different classes to handle errors
there is no longer a point for this class. This commit removes
it and changes ErrorManger to use Run directly.