Improved error handling in bootstrap startup
The error handler that is used for startup is not passed
on as the error handler for the app. The app will then
setup the error handler according to configuration and
the old is unregistered right before the new is added.
This should help with debugging problems in the startup
without messing up the normal error handler.