Added support for providing PSR loggers to eZDebug
The config ezp.log_mode controls how the eZDebug is configured
in regards to logging. 'psr' stops the internal logger and instead
creates a logger instance that is passed to eZDebug. 'ezdebug'
lets eZDebug log internally as before. 'disabled' disables both
the PSR and internal logging mechanism.
The loggers are configured by supplying a log factory in the
global variable $GLOBALS['ezpDebug']['loggerFactory'], this
takes a log-level string as parameter and returns a PSR logger
instance.

Note: This requires a patched eZ publish which acceps the log
configuration object.