Added proper handling of deprecation errors
Deprecations needs special care as one needs to turn them on
or off depending what code is run or the mode.

Added flag app.deprecation to control how deprecation errors are
treated. They can either be treated as an error, logged as a
an error or ignored. Defaults to logging for prod and error for
development.
This can be controlled at run-time by setting the env variable
`ERROR_DEPRECATION` to one of `error`, `log` or `ignore`.